[Refactor] Move descriptors
