+++ be/src/connector/jdbc_connector.cpp
@@ -20,6 +20,7 @@
 #include "exec/exec_node.h"
 #include "exec/jdbc_scanner.h"
 #include "exprs/expr.h"
+#include "runtime/descriptors_ext.h"
 #include "runtime/jdbc_driver_manager.h"
 #include "storage/chunk_helper.h"
