+++ be/src/connector/mysql_connector.cpp
@@ -17,6 +17,7 @@
 #include "column/chunk.h"
 #include "exprs/expr.h"
 #include "exprs/in_const_predicate.hpp"
+#include "runtime/descriptors_ext.h"
 #include "storage/chunk_helper.h"
 
 namespace starrocks::connector {
