+++ be/src/exec/hdfs_scanner/jni_scanner.h
@@ -23,6 +23,8 @@
 
 namespace starrocks {
 
+class HiveTableDescriptor;
+
 class JniScanner : public HdfsScanner {
 public:
     struct CreateOptions {
