+++ be/src/runtime/iceberg_table_sink.cpp
@@ -16,6 +16,7 @@
 
 #include "common/runtime_profile.h"
 #include "exprs/expr.h"
+#include "runtime/descriptors_ext.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks {
