+++ be/src/runtime/iceberg_table_sink.h
@@ -22,6 +22,7 @@
 namespace starrocks {
 
 class ExprContext;
+class IcebergTableDescriptor;
 struct TExprNode;
 
 class IcebergTableSink : public DataSink {
