+++ be/test/connector_sink/iceberg_table_sink_test.cpp
@@ -24,6 +24,7 @@
 #include "exec/pipeline/empty_set_operator.h"
 #include "exec/pipeline/fragment_context.h"
 #include "runtime/descriptor_helper.h"
+#include "runtime/descriptors_ext.h"
 #include "types/type_descriptor.h"
 
 namespace starrocks {
