+++ be/test/exec/hdfs_scanner/jni_scanner_test.cpp
@@ -17,6 +17,7 @@
 #include <gtest/gtest.h>
 
 #include "runtime/descriptor_helper.h"
+#include "runtime/descriptors_ext.h"
 #include "runtime/runtime_state.h"
 #include "util/thrift_util.h"
